Corrige migracion en tabla persona

+++ database/migrations/2018_03_27_201644_incluye_campos_google_id_y_facebook_id_en_persona.php
@@ -32,11 +32,15 @@ class IncluyeCamposGoogleIdYFacebookIdEnPersona extends Migration
      */
     public function down()
     {
-        Schema::table('Persona', function (Blueprint $table) {
-            $table->dropColumn(['google_id']);
-        });
-        Schema::table('Persona', function (Blueprint $table) {
-            $table->dropColumn(['facebook_id']);
-        });
+        if (Schema::hasColumn('Persona', 'google_id')) {
+            Schema::table('Persona', function (Blueprint $table) {
+                $table->dropColumn(['google_id']);
+            });
+        }
+        if (Schema::hasColumn('Persona', 'facebook_id')) {
+            Schema::table('Persona', function (Blueprint $table) {
+                $table->dropColumn(['facebook_id']);
+            });
+        }
     }
 }
